Korekcja do punktu 12. Usunięcie klasy.
