+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -2,6 +2,41 @@ package edu.kis.vh.nursery.list;
 
 public class IntLinkedList {
 
+    class Node {
+
+        private int value;
+        private Node prev;
+        private Node next;
+
+        Node(final int i) {
+            setValue(i);
+        }
+
+        int getValue() {
+            return value;
+        }
+
+        private void setValue(final int value) {
+            this.value = value;
+        }
+
+        Node getPrev() {
+            return prev;
+        }
+
+        void setPrev(final Node prev) {
+            this.prev = prev;
+        }
+
+        Node getNext() {
+            return next;
+        }
+
+        void setNext(final Node next) {
+            this.next = next;
+        }
+    }
+
     private static final int EMPTY_REPRESENTATION = -1;
     private Node last;
     private int i;
